+++ CHANGELOG.md
@@ -24,6 +24,7 @@
  - Fix SI1145 Input dependency
  - Fix Output/Measurement/Channel selection in Output Modules
  - Fix channel state at initialization of Python and Shell Command Outputs
+ - Fix function_status() error-handling
 
 ### Miscellaneous